Improve stability for OBS plugin
It is possible for destory to be called when injected frame is still due to display.
We use obs_source_get_frame to check if our frame is still being used (it also automatically removes the frame from the queue, so the loop should end after one cycle at most).